moving files into a component
