+++ imports/ui/components/externalApplications/helpModal.html
@@ -1,0 +1,48 @@
+<template name="AppHelpModal">
+    <div class="modal js-app-help-modal" tabindex="-1" role="dialog">
+        <div class="modal-dialog" role="document">
+            <div class="modal-content">
+                <div class="modal-header">
+                    <h5 class="modal-title">External applications help</h5>
+                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                        <span aria-hidden="true">&times;</span>
+                    </button>
+                </div>
+                <div class="modal-body">
+                    <p>Define an external link to be displayed on the Resources page</p>
+                    <dl>
+                        <dt>Name(required)</dt>
+                        <dd>Enter a name for you application</dd>
+                    
+                        <dt>URL(required)</dt>
+                        <dd>Enter a link for you application </dd>
+                        <dd>You can use the {{|name}} and {{|kind}} mustache templates as part of your URL</dd>
+                        <dd>Example: <code>https://company.example.com/{{|kind}}/{{|name}}</code> This would show up on the 
+                        Resources page as <code>https://company.example.com/Deployment/my-resource</code> if {{|kind}} matched <code>Deployment</code> and 
+                    {{|name}} matched <code>my-resource</code> </dd>
+                    
+                        <dt>Name Match(Regular expression)</dt>
+                        <dd>Example: <code>watch-keeper</code> would match resources named</dd>
+                        <dd>
+                            <code>watch-keeper</code><br>
+                            <code>watch-keeper-deploy</code><br>
+                            <code>watch-keeperapp</code><br>
+                        </dd>
+                        <dd>Leaving this blank will match any resource name</dd>
+
+                        <dt>Kind Match(Regular expression)</dt>
+                        <dd>Example: <code>Service</code> would match resources kinds named</dd>
+                        <dd>
+                            <code>Service</code><br>
+                            <code>ServiceAccount</code><br>
+                        </dd>
+                        <dd>Leaving this blank will match any resource type</dd>
+                    </dl>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-link" data-dismiss="modal">Close</button>
+                </div>
+            </div>
+        </div>
+    </div>
+</template>
