+++ imports/ui/pages/org/page.scss
@@ -14,10 +14,6 @@
 * limitations under the License.
 */
 
-.cursor-pointer {
-    cursor: pointer;
-}
-
 #editor {
     height: 500px;
 }
